Improve performance by introducing "numCandidates" parameter
Previously "numSamples" was increased from 100 to 1000 to make sure
that we get at least one ellipsoid candidate if many sampling attempts
failed. But this also slows down the ellipsoid fitting if we already
got enough candidates.

Introduce "numCandidates" restores the original performance, by
stopping the sampling as soon as enough candidates are found.